travail sur norminette choses simples + www nop sur les cast int car entraine seg fault

+++ prec.c
@@ -16,7 +16,7 @@ void	prec_s(t_elem *tmpa, char **str)
 {
 	char	*tmp;
 
-	if (SPREC.n != 0 && SPREC.n < (int)ft_strlen(*str))
+	if (SPREC.n != 0 && SPREC.n < ft_strlen(*str))
 	{
 		tmp = ft_strnew(SPREC.n + 1);
 		tmp = ft_strncpy(tmp, *str, SPREC.n);
@@ -24,20 +24,19 @@ void	prec_s(t_elem *tmpa, char **str)
 		*str = ft_strdup(tmp);
 		ft_strdel(&tmp);
 	}
-
 }
 
 void	prec_doux(t_elem *tmpa, char **str)
 {
-	char	*tmp;
-	size_t	i;
-	size_t	nb;
-	char	*add;
+	char		*tmp;
+	size_t		i;
+	size_t		nb;
+	char		*add;
 
 	i = 0;
-	if (SPREC.n != -1 && (SPREC.n > (int)ft_strlen(*str)))
+	if (SPREC.n != -1 && (SPREC.n > ft_strlen(*str)))
 	{
-	 	nb = SPREC.n - (int)ft_strlen(*str);
+		nb = SPREC.n - (int)ft_strlen(*str);
 		add = ft_strnew(nb);
 		while (i < nb)
 			add[i++] = '0';
